requirements.txt
astroid==2.4.2
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pyasn1==0.4.8
pycryptodome==3.9.8
pyjsbn-rsa==0.4.1
pylint==2.5.3
requests==2.24.0
rsa==4.6
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
urllib3==1.25.10
wrapt==1.12.1
requests==2.28.2
